tts_service/requirements.txt
chatterbox-tts
fastapi
uvicorn
python-multipart
torchaudio
torch
numpy
librosa
matplotlib
Pillow 